feat(di): Добавить юзкейсы в DI
