+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/app/di/DomainModule.kt
@@ -1,7 +1,13 @@
 package io.dimasla4ee.shoppinglist.app.di
 
+import io.dimasla4ee.shoppinglist.core.domain.interactor.CheckTokenUseCase
 import io.dimasla4ee.shoppinglist.core.domain.interactor.GetThemeInteractor
+import io.dimasla4ee.shoppinglist.core.domain.interactor.LoginUseCase
+import io.dimasla4ee.shoppinglist.core.domain.interactor.RecoverPasswordUseCase
+import io.dimasla4ee.shoppinglist.core.domain.interactor.RefreshTokenUseCase
+import io.dimasla4ee.shoppinglist.core.domain.interactor.RegisterUseCase
 import io.dimasla4ee.shoppinglist.core.domain.interactor.ToggleThemeInteractor
+import io.dimasla4ee.shoppinglist.core.domain.repository.AuthRepository
 import io.dimasla4ee.shoppinglist.core.domain.repository.SettingsRepository
 import io.dimasla4ee.shoppinglist.feature.shopping_lists.domain.ShoppingListsInteractor
 import io.dimasla4ee.shoppinglist.feature.shopping_lists.domain.ShoppingListsInteractorImpl
@@ -29,6 +35,36 @@ val domainModule = module {
     single<ShoppingListsInteractor> {
         ShoppingListsInteractorImpl(
             repository = get<ShoppingListsRepository>()
+        )
+    }
+
+    factory<RegisterUseCase> {
+        RegisterUseCase(
+            authRepository = get<AuthRepository>()
+        )
+    }
+
+    factory<LoginUseCase> {
+        LoginUseCase(
+            authRepository = get<AuthRepository>()
+        )
+    }
+
+    factory<RefreshTokenUseCase> {
+        RefreshTokenUseCase(
+            authRepository = get<AuthRepository>()
+        )
+    }
+
+    factory<RecoverPasswordUseCase> {
+        RecoverPasswordUseCase(
+            authRepository = get<AuthRepository>()
+        )
+    }
+
+    factory<CheckTokenUseCase> {
+        CheckTokenUseCase(
+            authRepository = get<AuthRepository>()
         )
     }
 }
