+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/domain/interactor/CheckTokenUseCase.kt
@@ -1,17 +1,16 @@
-package io.dimasla4ee.shoppinglist.core.data.repository
+package io.dimasla4ee.shoppinglist.core.domain.interactor
 
 import io.dimasla4ee.shoppinglist.core.domain.model.DomainResult
 import io.dimasla4ee.shoppinglist.core.domain.model.NetworkError
 import io.dimasla4ee.shoppinglist.core.domain.model.Response
 import io.dimasla4ee.shoppinglist.core.domain.repository.AuthRepository
 
-class LoginUseCase(
+class CheckTokenUseCase(
     private val authRepository: AuthRepository
 ) {
     suspend operator fun invoke(
-        email: String,
-        password: String
-    ): DomainResult<Response.UserAuthResponse, NetworkError> {
-        return authRepository.login(email, password)
+        accessToken: String
+    ): DomainResult<Response.CheckResponse, NetworkError> {
+        return authRepository.checkToken(accessToken)
     }
 }
