+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/domain/interactor/LoginUseCase.kt
@@ -1,17 +1,17 @@
-package io.dimasla4ee.shoppinglist.core.data.repository
+package io.dimasla4ee.shoppinglist.core.domain.interactor
 
 import io.dimasla4ee.shoppinglist.core.domain.model.DomainResult
 import io.dimasla4ee.shoppinglist.core.domain.model.NetworkError
 import io.dimasla4ee.shoppinglist.core.domain.model.Response
 import io.dimasla4ee.shoppinglist.core.domain.repository.AuthRepository
 
-class RegisterUseCase(
+class LoginUseCase(
     private val authRepository: AuthRepository
 ) {
     suspend operator fun invoke(
         email: String,
         password: String
-    ): DomainResult<Response.RegisterResponse, NetworkError> {
-        return authRepository.register(email, password)
+    ): DomainResult<Response.UserAuthResponse, NetworkError> {
+        return authRepository.login(email, password)
     }
 }
