+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/domain/interactor/RefreshTokenUseCase.kt
@@ -1,16 +1,16 @@
-package io.dimasla4ee.shoppinglist.core.data.repository
+package io.dimasla4ee.shoppinglist.core.domain.interactor
 
 import io.dimasla4ee.shoppinglist.core.domain.model.DomainResult
 import io.dimasla4ee.shoppinglist.core.domain.model.NetworkError
 import io.dimasla4ee.shoppinglist.core.domain.model.Response
 import io.dimasla4ee.shoppinglist.core.domain.repository.AuthRepository
 
-class CheckTokenUseCase(
+class RefreshTokenUseCase(
     private val authRepository: AuthRepository
 ) {
     suspend operator fun invoke(
-        accessToken: String
-    ): DomainResult<Response.CheckResponse, NetworkError> {
-        return authRepository.checkToken(accessToken)
+        refreshToken: String
+    ): DomainResult<Response.RefreshTokenResponse, NetworkError> {
+        return authRepository.refresh(refreshToken)
     }
 }
