+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/domain/interactor/RegisterUseCase.kt
@@ -1,16 +1,17 @@
-package io.dimasla4ee.shoppinglist.core.data.repository
+package io.dimasla4ee.shoppinglist.core.domain.interactor
 
 import io.dimasla4ee.shoppinglist.core.domain.model.DomainResult
 import io.dimasla4ee.shoppinglist.core.domain.model.NetworkError
 import io.dimasla4ee.shoppinglist.core.domain.model.Response
 import io.dimasla4ee.shoppinglist.core.domain.repository.AuthRepository
 
-class RefreshTokenUseCase(
+class RegisterUseCase(
     private val authRepository: AuthRepository
 ) {
     suspend operator fun invoke(
-        refreshToken: String
-    ): DomainResult<Response.RefreshTokenResponse, NetworkError> {
-        return authRepository.refresh(refreshToken)
+        email: String,
+        password: String
+    ): DomainResult<Response.RegisterResponse, NetworkError> {
+        return authRepository.register(email, password)
     }
 }
